addon: i/o types
I placed these in the NSPS encoder as this made most sense to keep the right types together with the right ios. todo: make this optional to allow still using the old model as well. that may be tricky tho since they imply different numbers of params in the function type signatures...?